requirements.txt
ipython>=7.5.0
pygame>=1.9
pytest>=5.1
pytest-repeat>=0.8.0
pyyaml>=5.1.2
dill==0.3.0
numpy==1.17.0
matplotlib==3.1.1
scipy==1.3.1
attrs==19.1.0
